refactor(prj): separate proxy util to core module
