+++ core/go.mod
@@ -7,10 +7,12 @@ require (
 	github.com/gabriel-vasile/mimetype v1.4.4
 	github.com/go-faster/errors v0.7.1
 	github.com/gotd/td v0.102.0
+	github.com/iyear/connectproxy v0.1.1
 	github.com/yapingcat/gomedia v0.0.0-20240601043430-920523f8e5c7
 	go.uber.org/atomic v1.11.0
 	go.uber.org/multierr v1.11.0
 	go.uber.org/zap v1.27.0
+	golang.org/x/net v0.25.0
 	golang.org/x/sync v0.7.0
 )
 
@@ -24,7 +26,6 @@ require (
 	go.opentelemetry.io/otel v1.26.0 // indirect
 	go.opentelemetry.io/otel/trace v1.26.0 // indirect
 	golang.org/x/crypto v0.23.0 // indirect
-	golang.org/x/net v0.25.0 // indirect
 	golang.org/x/sys v0.20.0 // indirect
 	nhooyr.io/websocket v1.8.11 // indirect
 	rsc.io/qr v0.2.0 // indirect
